Add support for using ifconfig and route's '-j' option
On systems which support it, this allows the container's network to be
setup without relying on the ifconfig and route binaries to be present
in the network namespace jail.

Without this support, two jails are needed, one to own the vnet and one
to run the container. The vnet owning jail must either container
ifconfig and route binaries or must share the host's root path.